pyproject.toml
[project]
name = "hermes"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "reflex>=0.8.0",
    "pyserial>=3.5",
    "plotly>=5.0",
    "pandas>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hermes"]
